Add more validation to books. Fix POST /books. Correct Joi-validation calls.
Validation of objectIds, not just strings. Catch a TyepError that can surface on res.ops after a insert that wasn't succesful for some reason.

Joi-validation calls where misunderstood. The object returned form `Joi.validate` has a field `error` which is null if there are no errors. The call does not return a boolean as was understood before.
This is corrected in both crowd and book-model